Add broadcast to the coords frames data set. Make sure we use long int when broadcasting dataset size.

+++ src/DataSet_Mat3x3.cpp
@@ -53,8 +53,8 @@ int DataSet_Mat3x3::Sync(size_t total, std::vector<int> const& rank_frames,
 int DataSet_Mat3x3::Bcast(Parallel::Comm const& commIn) {
   if (commIn.Size() == 1) return 0;
   // Assume all data is currently on the master process.
-  int totalSize = Size();
-  int err = commIn.MasterBcast( &totalSize, 1, MPI_INT );
+  long int totalSize = Size();
+  int err = commIn.MasterBcast( &totalSize, 1, MPI_LONG );
   if (!commIn.Master()) {
     //rprintf("DEBUG: Resizing array to %i\n", totalSize);
     Resize( totalSize );
